Current work on fixing localisation bug

+++ embasi/roothan_hall_eigensolver_scalapack.py
@@ -106,7 +106,7 @@ def hamiltonian_eigensolv_parallel(hamiltonian, overlap, nelec, nspins=1, nkpts=
     # compare eigenvalues
     occ_mat = {}
     if nspins > 1:
-        remaining_electrons = int(round(nelec))
+        remaining_electrons = round(nelec)
         alpha_nelecs = 0
         beta_nelecs = 0
         occ_mat[(0,0)] = np.zeros(np.size(evals[(0,0)]))
@@ -123,8 +123,8 @@ def hamiltonian_eigensolv_parallel(hamiltonian, overlap, nelec, nspins=1, nkpts=
             remaining_electrons += -1
     else:
         occ_mat[(0,0)] = np.zeros(np.size(evals[(0,0)]))
-        occ_mat[(0,0)][:int(round(nelec/2))] = 2.0
-        
+        occ_mat[(0,0)][:round(nelec/2)] = 2.0
+
     evecs = SpinKpointArray(evecs, nspins, nkpts)
     evals = SpinKpointArray(evals, nspins, nkpts)
     occ_mat = SpinKpointArray(occ_mat, nspins, nkpts)
